Add route: /update_strategy_config

+++ server_app/src/strategy/check_points.cpp
@@ -2,6 +2,6 @@
 
 CheckPoints::CheckPoints(const std::string symbol, double volumn, double move_price)
 {
-    std::string collection_name = symbol + "_" + std::to_string(volumn) + "_" + std::to_string(move_price);
+    std::string collection_name = symbol + "_" + std::to_string((size_t)volumn) + "_" + std::to_string((size_t)move_price);
     m_checkpoint_list = DataModel::get_data_model_map(STRATEGY_DB_NAME, collection_name, "checkpoint_id");
 }
